+++ core/core-api/src/main/java/cmc/mellyserver/controller/group/dto/request/GroupCreateRequest.java
@@ -1,13 +1,12 @@
 package cmc.mellyserver.controller.group.dto.request;
 
 import cmc.mellyserver.dbcore.group.enums.GroupType;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Data
