+++ core/core-api/src/main/java/cmc/mellyserver/controller/group/dto/request/GroupUpdateRequest.java
@@ -1,11 +1,10 @@
 package cmc.mellyserver.controller.group.dto.request;
 
 import cmc.mellyserver.dbcore.group.enums.GroupType;
+import jakarta.validation.constraints.Size;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.Size;
 
 @Data
 @NoArgsConstructor
