+++ core/core-api/src/main/java/cmc/mellyserver/domain/group/GroupService.java
@@ -1,6 +1,8 @@
 package cmc.mellyserver.domain.group;
 
 
+import cmc.mellyserver.common.aop.lock.annotation.DistributedLock;
+import cmc.mellyserver.common.aop.lock.annotation.OptimisticLock;
 import cmc.mellyserver.dbcore.group.GroupAndUser;
 import cmc.mellyserver.dbcore.group.GroupAndUserRepository;
 import cmc.mellyserver.dbcore.group.GroupRepository;
@@ -13,8 +15,6 @@ import cmc.mellyserver.domain.group.dto.response.GroupListLoginUserParticipatedR
 import cmc.mellyserver.domain.group.query.UserGroupQueryRepository;
 import cmc.mellyserver.domain.group.query.dto.GroupDetailResponseDto;
 import cmc.mellyserver.domain.group.query.dto.GroupLoginUserParticipatedResponseDto;
-import cmc.mellyserver.common.aop.lock.annotation.DistributedLock;
-import cmc.mellyserver.common.aop.lock.annotation.OptimisticLock;
 import cmc.mellyserver.support.exception.BusinessException;
 import cmc.mellyserver.support.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
