+++ core/core-api/src/main/java/cmc/mellyserver/domain/group/query/dto/GroupDetailResponseDto.java
@@ -1,4 +1,4 @@
-package cmc.mellyserver.dbcore.group.query.dto;
+package cmc.mellyserver.domain.group.query.dto;
 
 import cmc.mellyserver.dbcore.group.UserGroup;
 import cmc.mellyserver.dbcore.group.enums.GroupType;
