+++ core/core-api/src/main/java/cmc/mellyserver/domain/group/query/dto/GroupLoginUserParticipatedResponseDto.java
@@ -1,4 +1,4 @@
-package cmc.mellyserver.dbcore.group.query.dto;
+package cmc.mellyserver.domain.group.query.dto;
 
 import cmc.mellyserver.dbcore.group.GroupAndUser;
 import cmc.mellyserver.dbcore.group.enums.GroupType;
