+++ core/core-api/src/main/java/cmc/mellyserver/domain/memory/query/dto/FindPlaceInfoByMemoryNameResponseDto.java
@@ -1,4 +1,4 @@
-package cmc.mellyserver.dbcore.memory.query.dto;
+package cmc.mellyserver.domain.memory.query.dto;
 
 import lombok.Builder;
 import lombok.Data;
