+++ core/core-api/src/main/java/cmc/mellyserver/domain/scrap/query/dto/PlaceScrapCountResponseDto.java
@@ -1,4 +1,4 @@
-package cmc.mellyserver.dbcore.scrap.query.dto;
+package cmc.mellyserver.domain.scrap.query.dto;
 
 import cmc.mellyserver.dbcore.scrap.enums.ScrapType;
 import lombok.AllArgsConstructor;
