+++ storage/db-core/build.gradle
@@ -1,7 +1,9 @@
 dependencies {
 
     api 'org.springframework.boot:spring-boot-starter-data-jpa'
+
     runtimeOnly 'com.mysql:mysql-connector-j'
     runtimeOnly 'com.h2database:h2'
 }
+
